+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+## Added
+- precision command to set the precision of Real number calculations globally (for the CLI)
+
 
 ## [v2.1.0]